Refactored class and fixed bug in TileMap class

+++ include/TileSet.h
@@ -14,7 +14,7 @@ private:
 
 public:
     TileSet(GameObject &associated, int tileWidth, int tileHeight, std::string file);
-    void RenderTile(unsigned index, float x, float y);
+    void RenderTile(unsigned index, int x, int y);
     int GetTileWidth();
     int GetTileHeight();
 };
